requirements.txt
omegaconf
pillow
imageio-ffmpeg
opencv-python
scikit-image
einops
tqdm
torch==2.4.1
torchvision==0.19.1
transformers==4.45.2
ultralytics
gradio
pycocotools==2.0.8
paddlepaddle-gpu
paddleocr>=2.0.1
openai>=1.54.0
playwright>=1.50.0
python-dotenv
requests